Proper content-script and infobox without background-page
The infobox.html page is now a division which is injected into the
page with a double click and removed with a single click. For this
to work the manifest file was altered. Furthermore the background
page and therefore the message passing system has been removed
since there is no obvious use for it.