Fix issues concerning the zip file and the Store
The extension did not seem to work when downloaded through the Google
Chrome Web Store. Since the extension works just fine when either
installing the .crx or the pure files itself, it must be due to a false
zip file. Therefore filesync the zip, plus omit the top-level directory.
This solves the issue.

Signed-off-by: Gordian Edenhofer <gordian.edenhofer@gmail.com>